Title: 자주 등장한 top k 숫자, Time: 865ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240405/자주 등장한 top k 숫자/top-k-frequent-elements.cpp
@@ -5,8 +5,8 @@ using namespace std;
 
 int main() {
 	int n, k; cin >> n >> k;
-	vector<int> v(n);
-	unordered_map<int, int> map;
+	vector<long long> v(n);
+	unordered_map<long long, int> map;
 
 	for (int i = 0; i < n; i++) {
 		cin >> v[i];
@@ -15,7 +15,7 @@ int main() {
 
 	while (k--) {
 		int max_count = 0;
-		int max_elemt = 0;
+		long long max_elemt = 0;
 
 		for (int i = n - 1; i >= 0; i--) {
 			if (map[v[i]] > max_count) {
